python/requirements.txt
PyHamcrest>=2.1.0
PyYaml>=6.0.2
behave>=1.2.6
docker>=7.1.0
mockseries>=0.3.1
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation>=0.48b0
opentelemetry-instrumentation-openai>=0.33.1
opentelemetry-exporter-prometheus-remote-write>=0.48b0
opentelemetry-semantic-conventions>=0.48b0
opentelemetry-semantic-conventions-ai>=0.4.1
pillow>=10.4.0
python-dotenv>=1.0.1
requests>=2.32.3
regex>=2024.9.11
Flask>=3.0.3
python-dateutil>=2.9.0.post0
pyjwt>=2.9.0